kth/ 전체적 수정

+++ src/main/java/com/workie/easy/work/model/dto/VacationInfo.java
@@ -15,5 +15,5 @@ public class VacationInfo {
 
 	private int empNo;
 	private int allDays;
-	private int leftDays;
+	private double leftDays;
 }
